[WIP] getting the PCIe0 for wifi & bt running
... and other sad stories

bluetooth is working (yay), and the pcie controller does say hi, but I
can't for the love of it discover the wifi chip on the bus.

rn the init sequence of the phy is 1:1 the one from the downstream cape
device, bluetooth works (so chip is prob. powered) but yeah. still no
see

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>